+++ nemo/lightning/pytorch/utils.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import lightning.pytorch as pl
 import torch
 
 
@@ -55,3 +56,14 @@ def dtype_from_hf(config):
         return dtype_from_str(torch_dtype)
     else:
         raise ValueError("torch_dtype is not of type str/torch.dtype")
+
+
+def is_trainer_attached(model: pl.LightningModule):
+    """
+    Returns true if trainer is attached to a model
+    """
+    try:
+        trainer = model.trainer
+        return True
+    except (AttributeError, RuntimeError):
+        return False
